+++ changelog.md
@@ -2,6 +2,13 @@
 
 ## UNRELEASED
 
+**Added:**
+- bpk-component-image:
+  - Added `BpkImage` component
+  - Added `BpkBackgroundImage` component which can be used to create a container with a background image
+  - Added `withLazyLoading` HOC which allows `BpkImage` or `BpkBackgroundImage` to be loaded only when in view
+  - Added `withLoadingBehavior` HOC which allows `BpkImage` or `BpkBackgroundImage` to be shown only once loaded, displaying a spinner until that point
+
 **Fixed:**
 - bpk-component-button:
   - Fixed SyntheticEvent not being passed to the onClick handler.